Clean up generation of typescript types from graphql
This changes the graphql generation such that
* the generated types are located in `common/graphql/types.ts`
* the imported code does not involve the resolvers
* there are no imports from `common` to `public` or `server`
* the root schema is shared between client and server
* a `public/graphql` file collects all future client-side schemas
