+++ x-pack/package.json
@@ -64,8 +64,8 @@
     "enzyme-adapter-react-16": "^1.1.1",
     "enzyme-to-json": "3.3.1",
     "expect.js": "0.3.1",
-    "graphql-code-generator": "^0.9.1",
-    "graphql-codegen-typescript-template": "^0.9.1",
+    "graphql-code-generator": "^0.10.1",
+    "graphql-codegen-typescript-template": "^0.10.1",
     "gulp": "3.9.1",
     "gulp-load-plugins": "1.2.0",
     "gulp-mocha": "2.2.0",
