+++ x-pack/plugins/infra/common/graphql/root/index.ts
@@ -4,6 +4,4 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import rootSchema from './root.gql_schema';
-
-export { rootSchema };
+export { rootSchema } from './schema.gql';
