+++ x-pack/plugins/infra/server/graphql/fields/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export { createFieldResolvers } from './resolvers';
+export { fieldsSchema } from './schema.gql';
