+++ x-pack/plugins/infra/server/graphql/nodes/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export { createNodeResolvers } from './resolvers';
+export { nodesSchema } from './schema.gql';
