+++ x-pack/plugins/infra/server/lib/adapters/nodes/elasticsearch_nodes_adapter.ts
@@ -14,7 +14,7 @@ import {
 } from '../../infra_types';
 
 import { DOMAIN_TO_FIELD } from '../../../../common/constants';
-import { InfraResponse } from '../../../../common/types';
+import { InfraResponse } from '../../../../common/graphql/types';
 import { calculateCardinalityOfNodeField } from './lib/calculate_cardinality';
 import { createPartitionBodies } from './lib/create_partition_bodies';
 import { processNodes } from './lib/process_nodes';
