+++ x-pack/plugins/infra/server/lib/adapters/nodes/lib/convert_nodes_response_to_groups.ts
@@ -11,7 +11,7 @@ import { findOrCreateGroupWithNodes, findOrCreateGroupWithSubGroups } from './fi
 import { InfraBucket, InfraNodeRequestOptions } from '../../../infra_types';
 
 import { set } from 'lodash';
-import { InfraGroup, InfraGroupBy } from '../../../../../common/types';
+import { InfraGroup, InfraGroupBy } from '../../../../../common/graphql/types';
 
 export function convertNodesResponseToGroups(
   options: InfraNodeRequestOptions,
