+++ x-pack/plugins/infra/server/lib/adapters/nodes/lib/process_nodes.ts
@@ -5,7 +5,7 @@
  */
 
 import { set } from 'lodash';
-import { InfraResponse } from '../../../../../common/types';
+import { InfraResponse } from '../../../../../common/graphql/types';
 import {
   InfraBucket,
   InfraNode,
